+++ src/index.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Matterbridge, MatterbridgeDevice, MatterbridgeEndpoint, PlatformConfig } from 'matterbridge';
+import { Matterbridge, MatterbridgeEndpoint, PlatformConfig } from 'matterbridge';
 import { AnsiLogger, LogLevel } from 'matterbridge/logger';
 import { ZigbeePlatform } from './platform.js';
 import initializePlugin from './index';
@@ -26,7 +26,7 @@ describe('initializePlugin', () => {
       matterbridgeDirectory: './jest/matterbridge',
       matterbridgePluginDirectory: './jest/plugins',
       systemInformation: { ipv4Address: undefined },
-      matterbridgeVersion: '1.7.3',
+      matterbridgeVersion: '2.1.0',
       getDevices: jest.fn(() => {
         // console.log('getDevices called');
         return [];
@@ -35,21 +35,12 @@ describe('initializePlugin', () => {
         // console.log('getDevices called');
         return [];
       }),
-      addBridgedDevice: jest.fn(async (pluginName: string, device: MatterbridgeDevice) => {
-        // console.log('addBridgedDevice called');
-      }),
       addBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
         // console.log('addBridgedEndpoint called');
         // await aggregator.add(device);
       }),
-      removeBridgedDevice: jest.fn(async (pluginName: string, device: MatterbridgeDevice) => {
-        // console.log('removeBridgedDevice called');
-      }),
       removeBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
         // console.log('removeBridgedEndpoint called');
-      }),
-      removeAllBridgedDevices: jest.fn(async (pluginName: string) => {
-        // console.log('removeAllBridgedDevices called');
       }),
       removeAllBridgedEndpoints: jest.fn(async (pluginName: string) => {
         // console.log('removeAllBridgedEndpoints called');
@@ -124,7 +115,7 @@ describe('initializePlugin', () => {
 
   it('should return an instance of TestPlatform', () => {
     const result = initializePlugin(mockMatterbridge, mockLog, mockConfig);
-
     expect(result).toBeInstanceOf(ZigbeePlatform);
+    result.onShutdown();
   });
 });